move gif processing to own file
