+++ src/utils/gif.rs
@@ -1,7 +1,23 @@
-use std::{io::Write, path::{PathBuf, Path}, process::Command};
+use std::{
+    fs::File,
+    io::Write,
+    path::{Path, PathBuf},
+    process::Command,
+};
 
 use anyhow::Result;
+use colored::Colorize;
+use image::{
+    gif::{GifDecoder, GifEncoder},
+    AnimationDecoder, GenericImage, ImageDecoder, RgbaImage,
+};
 use log::{info, warn};
+use rayon::iter::{IntoParallelRefMutIterator, ParallelIterator};
+use rusttype::Font;
+
+use crate::utils::image::{SetUp, TextImage};
+
+use super::args::Cli;
 
 pub struct Gifsicle {
     exe: PathBuf,
@@ -67,4 +83,52 @@ impl Gifsicle {
         info!("The optimization will be complete when the terminal window closes.");
         Ok(())
     }
+}
+
+pub fn process_gif(
+    gif: File,
+    font: Font<'static>,
+    text: &str,
+    out_path: &Path,
+    name: &str,
+    cli: &Cli,
+) -> Result<(), anyhow::Error> {
+    let decoder = GifDecoder::new(gif)?;
+    let (gif_w, gif_h) = decoder.dimensions();
+    let init = SetUp::init(font).with_dimensions(gif_w, gif_h);
+    info!("Creating caption image...");
+    let image = TextImage::new(init, text).render()?;
+    info!("{}", "Caption image created!".green());
+    let mut frames = decoder.into_frames().collect_frames()?;
+    info!("{}", "Rendering GIF...".blue());
+    frames.par_iter_mut().for_each(|f| {
+        let f = f.buffer_mut();
+        let mut buffer = RgbaImage::new(gif_w, gif_h + image.height());
+
+        buffer
+            .copy_from(&image, 0, 0)
+            .expect("could not copy buffer");
+
+        buffer
+            .copy_from(f, 0, image.height())
+            .expect("could not copy buffer");
+
+        *f = buffer;
+    });
+    let file_out = File::create(&out_path.join(&name))?;
+    let file_out_path = out_path.join(&name);
+    let mut encoder = GifEncoder::new_with_speed(&file_out, 30);
+    encoder.set_repeat(image::gif::Repeat::Infinite)?;
+    encoder.encode_frames(frames)?;
+    info!(
+        "GIF: {} {} at {}",
+        &name,
+        "generated".green(),
+        out_path.to_str().expect("invalid output path"),
+    );
+    let opt = cli.opt_level().map(std::borrow::ToOwned::to_owned);
+    let lossy = cli.lossy();
+    let reduce = cli.reduce();
+    Gifsicle::init()?.run(opt, lossy, reduce, &file_out_path)?;
+    Ok(())
 }
